+++ server/Services/Product/API/DTOs/CreateOrderDto.cs
@@ -9,7 +9,7 @@ public class CreateOrderDto
     public string CartId { get; set; } = string.Empty;
 
     [Required]
-    public int DeliveryMethodId { get; set; }
+    public long DeliveryMethodId { get; set; }
 
     [Required]
     public ShippingAddress ShippingAddress { get; set; } = null!;
