+++ server/Services/Product/API/DTOs/ProductAttributeValueIndexDto.cs
@@ -2,28 +2,28 @@
 {
     public class ProductAttributeValueIndexDto
     {
-        public int Id { get; set; }
+        public long Id { get; set; }
         public string? OptionText { get; set; }
-        public int VariantId { get; set; }
-        public int AttributeId { get; set; }
+        public long VariantId { get; set; }
+        public long AttributeId { get; set; }
         public AttributeOptionDto? Option { get; set; }
     }
 
     public class CreateProductAttributeValueIndexDto
     {
         public string? OptionText { get; set; }
-        public int VariantId { get; set; }
-        public int AttributeId { get; set; }
-        public int OptionId { get; set; }
+        public long VariantId { get; set; }
+        public long AttributeId { get; set; }
+        public long OptionId { get; set; }
     }
 
     public class UpdateProductAttributeValueIndexDto
     {
-        public int Id { get; set; }
+        public long Id { get; set; }
 
         public string? OptionText { get; set; }
-        public int VariantId { get; set; }
-        public int AttributeId { get; set; }
-        public int OptionId { get; set; }
+        public long VariantId { get; set; }
+        public long AttributeId { get; set; }
+        public long OptionId { get; set; }
     }
 }
