+++ server/Services/Product/Core/Entities/ProductAttributeNavigator.cs
@@ -3,9 +3,9 @@ namespace Core.Entities;
 
 public class ProductAttributeNavigator
 {
-    public int ProductId { get; set; }
+    public long ProductId { get; set; }
     public Product Product { get; set; } = null!;
 
-    public int ProductAttributeId { get; set; }
+    public long ProductAttributeId { get; set; }
     public ProductAttribute ProductAttribute { get; set; } = null!;
 }
