+++ server/Services/Product/Core/Interfaces/IGenericRepository.cs
@@ -4,7 +4,7 @@ namespace Core.Interfaces;
 
 public interface IGenericRepository<T> where T : BaseEntity
 {
-    Task<T?> GetByIdAsync(int id);
+    Task<T?> GetByIdAsync(long id);
     Task<IReadOnlyList<T>> ListAllAsync();
     Task<T?> GetEntityWithSpec(ISpecification<T> spec);
     Task<IReadOnlyList<T>> ListAsync(ISpecification<T> spec);
@@ -13,6 +13,6 @@ public interface IGenericRepository<T> where T : BaseEntity
     void Add(T entity);
     void Update(T entity);
     void Remove(T entity);
-    bool Exists(int id);
+    bool Exists(long id);
     Task<int> CountAsync(ISpecification<T> spec);
 }
